player/requirements.txt
atomicwrites>=1.3.0
attrs>=19.1.0
importlib-metadata>=0.17
more-itertools>=7.0.0
pluggy>=0.12.0
py>=1.8.0
pytest>=4.5.0
six>=1.12.0
wcwidth>=0.1.7
zipp>=0.5.1
pylint
pyyaml